# docs/architecture-pack/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Farmmap API
  version: 1.0.0
  description: >
    Farmmap public and authenticated API covering all Must-Have features (F1–F16).
    All endpoints require a Supabase JWT Bearer token where marked. directory_id
    is resolved server-side from the Host header; it is never a client parameter
    on public endpoints.
  contact:
    name: Farmmap Engineering
    url: https://farmmap.co.uk
servers:
  - url: https://farmmap.co.uk/api
    description: Production
  - url: https://staging.farmmap.co.uk/api
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from auth.getSession()
  parameters:
    ListingSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        example: hargreaves-farm-shop
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Listing not found
        details:
          type: object
          additionalProperties: true
    ListingType:
      type: string
      enum: [farm_shop, honesty_box, farm_gate_stall, roadside_stand]
    ListingStatus:
      type: string
      enum: [pending, active, claimed, suspended]
    SubscriptionTier:
      type: string
      enum: [free, bronze, silver, gold]
    OpeningHoursDay:
      type: object
      properties:
        open:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: '09:00'
        close:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: '17:00'
        closed:
          type: boolean
    OpeningHours:
      type: object
      properties:
        monday:
          $ref: '#/components/schemas/OpeningHoursDay'
        tuesday:
          $ref: '#/components/schemas/OpeningHoursDay'
        wednesday:
          $ref: '#/components/schemas/OpeningHoursDay'
        thursday:
          $ref: '#/components/schemas/OpeningHoursDay'
        friday:
          $ref: '#/components/schemas/OpeningHoursDay'
        saturday:
          $ref: '#/components/schemas/OpeningHoursDay'
        sunday:
          $ref: '#/components/schemas/OpeningHoursDay'
        notes:
          type: string
          description: Free text for irregular patterns, e.g. "Tuesday and Friday only"
          maxLength: 500
        bank_holidays_closed:
          type: boolean
    GeoPoint:
      type: object
      required: [lon, lat]
      properties:
        lon:
          type: number
          format: double
          minimum: -180
          maximum: 180
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
    ListingMeta:
      type: object
      description: >
        Listing-type-specific metadata. For honesty_box:
        payment_methods, currently_stocked, stocked_updated_at, location_description.
      additionalProperties: true
    ListingSummary:
      type: object
      required: [id, slug, name, listing_type, status, tier, town, county, country, location]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        listing_type:
          $ref: '#/components/schemas/ListingType'
        status:
          $ref: '#/components/schemas/ListingStatus'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        town:
          type: string
        county:
          type: string
        country:
          type: string
        location:
          $ref: '#/components/schemas/GeoPoint'
        review_rating_avg:
          type: number
          format: float
          nullable: true
        review_count:
          type: integer
        hero_photo_url:
          type: string
          format: uri
          nullable: true
    ListingDetail:
      allOf:
        - $ref: '#/components/schemas/ListingSummary'
        - type: object
          properties:
            address_line1:
              type: string
            address_line2:
              type: string
            postcode:
              type: string
            eircode:
              type: string
            phone:
              type: string
            website:
              type: string
              format: uri
            description:
              type: string
            opening_hours:
              $ref: '#/components/schemas/OpeningHours'
            temporarily_closed:
              type: boolean
            listing_type_meta:
              $ref: '#/components/schemas/ListingMeta'
            photos:
              type: array
              items:
                $ref: '#/components/schemas/ListingPhoto'
            geocoding_approximate:
              type: boolean
    ListingPhoto:
      type: object
      required: [id, url, thumbnail_url, alt_text]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        alt_text:
          type: string
        display_order:
          type: integer
...
    Review:
      type: object
      required: [id, rating, body, created_at]
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        body:
          type: string
          minLength: 20
          maxLength: 1000
        owner_response:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /listings/{slug}:
    get:
      operationId: getListing
      summary: Get listing detail
      description: >
        Returns full listing detail including photos (approved only), opening hours,
        and listing-type-specific metadata. No authentication required.
      tags: [Listings]
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateListing
      summary: Update listing fields
      description: >
        Updates listing fields. Requires owner or manager role for the listing.
        Photos are managed separately via /photos endpoints.
        All text fields take effect immediately; photo changes require moderation.
      tags: [Listings]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                description:
                  type: string
                  maxLength: 5000
                address_line1:
                  type: string
                address_line2:
                  type: string
                town:
                  type: string
                county:
                  type: string
                postcode:
                  type: string
                eircode:
                  type: string
                phone:
                  type: string
                website:
                  type: string
                  format: uri
                opening_hours:
                  $ref: '#/components/schemas/OpeningHours'
                temporarily_closed:
                  type: boolean
                listing_type_meta:
                  $ref: '#/components/schemas/ListingMeta'
      responses:
        '200':
          description: Updated listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ============================================================
  # REVIEWS (F5)
  # ============================================================
  /listings/{slug}/reviews:
    get:
      operationId: listReviews
      summary: Get paginated approved reviews for a listing
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated approved reviews
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  aggregate_rating:
                    type: number
                    nullable: true
                    description: Null if fewer than 3 approved reviews
                  review_count:
                    type: integer

    post:
      operationId: submitReview
      summary: Submit a consumer review
      description: >
        Authenticated consumer submits a star rating and text review.
        Review enters moderation queue; not visible until approved.
        One review per consumer per listing enforced.
      tags: [Reviews]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating, body]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                body:
                  type: string
                  minLength: 20
                  maxLength: 1000
      responses:
        '201':
          description: Review submitted and pending moderation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: Your review has been submitted and is pending moderation.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Consumer has already reviewed this listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ============================================================
  # ORDERING WAITLIST (F6)
  # ============================================================
...
  /events:
    post:
      operationId: trackEvent
      summary: Record a server-side analytics event
      description: >
        Internal endpoint called by Next.js RSC and API handlers to record
        business events (listing_page_view, map_pin_click, etc.) in the
        analytics_events table. No PII stored. Rate limited per IP.
      tags: [Analytics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                event_type:
                  type: string
                  enum:
                    - listing_page_view
                    - listing_pin_click
                    - claim_cta_click
                    - subscription_upgrade_initiated
                    - subscription_upgrade_completed
                    - waitlist_signup
                    - review_submission
                    - enquiry_form_submission
                listing_id:
                  type: string
                  format: uuid
                  nullable: true
                source:
                  type: string
                  enum: [organic, direct, social, referral]
                metadata:
                  type: object
                  description: Additional context with no PII
                  additionalProperties: true
      responses:
        '204':
          description: Event recorded
        '429':
          $ref: '#/components/responses/TooManyRequests'
...